+++ VSC Extension/config/messages.js
@@ -2,14 +2,14 @@ module.exports = {
     //Informations
     argonRunning: 'Argon is now running!',
     argonStopped: 'Argon stopped!',
-    databaseUpdated: 'Successfully updated class database',
+    databaseUpdated: 'Successfully updated class database!',
     databaseUpToDate: 'Class database is already up to date!',
     
     //Warnings
     notRunning : 'Argon is not running!',
     alreadyRunning: 'Argon is already running!',
     openWorkspace: 'Please open workspace first!',
-    updatingDatabase: 'Please wait database is updating!',
+    updatingDatabase: 'Please wait, the database is being updated!',
     outdatedVersion: 'You are running outdated version, please install latest update!',
 
     //Errors